Fixed a warning in the timeo test
Fixed a couple of mixed signed/unsigned comparison which in conjunction
with -Werror prevented compilation of the test.

+++ tests/timeo.cpp
@@ -66,7 +66,7 @@ int XS_TEST_MAIN ()
     assert (rc == -1);
     assert (xs_errno () == EAGAIN);
     unsigned long elapsed = xs_stopwatch_stop (watch) / 1000;
-    time_assert (elapsed, timeout);
+    time_assert (elapsed, (unsigned long) timeout);
 
     //  Check whether connection during the wait doesn't distort the timeout.
     timeout = 2000;
@@ -79,7 +79,7 @@ int XS_TEST_MAIN ()
     assert (rc == -1);
     assert (xs_errno () == EAGAIN);
     elapsed = xs_stopwatch_stop (watch) / 1000;
-    time_assert (elapsed, timeout);
+    time_assert (elapsed, (unsigned long) timeout);
     thread_join (thread);
 
     //  Check that timeouts don't break normal message transfer.
